requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
Pillow==9.5.0
numpy==1.26.4
requests
PyYAML
